+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 0.4.6
+- Support pegasus project debugger.
 - feat: add quick entries in tree view.
 
 ## 0.4.5